Move policy related funcs from utils to policy module
