+++ code/modules/fakegres/exsiting.go
@@ -3,19 +3,19 @@ package fakegres
 import (
 	"fmt"
 
+	"github.com/welibekov/grantmaster/modules/policy"
 	"github.com/welibekov/grantmaster/modules/policy/types"
-	"github.com/welibekov/grantmaster/modules/utils"
 )
 
 // readExistingPolicies reads all YAML files in the specified directory and unmarshals them into Policy structs.
 func (f *Fakegres) readExistingPolicies() (map[string][]string, error) {
 	policiesMap := make(map[string][]string)
 
-	policies, err := utils.ReadPoliciesFromDirectory(f.rootDir,
+	policies, err := policy.ReadPoliciesFromDirectory(f.rootDir,
 		func(path string) ([]types.Policy, error) {
 			policies := []types.Policy{}
 
-			policy, err := utils.ReadPolicy[types.Policy](path)
+			policy, err := policy.ReadPolicy[types.Policy](path)
 			if err != nil {
 				return policies, err
 			}
